chore: address dead code and code-review nits
Remove unused members and locals (async_response_handler result_message_/
elapsed_millis_, DigitalInputState triggered_, BaseAnalogReader output_scale_,
base_command_handler sz). Fix narrowing of kInvalidDouble to float in ENUVector
and AttitudeVector, ENUVector canConvertFromJson field type, missing override on
VoltageMultiplier::set, namespace scope of kMaxAnalogOutput, bool return values
in HTTPDigestAuthenticator::authenticate_request, Content-Type comparison by
value, stale include-guard and namespace comments, redundant LEDPattern member
init, and per-instance blip state in BaseBlinker.

+++ src/sensesp/types/position.cpp
@@ -33,7 +33,7 @@ bool canConvertFromJson(JsonVariantConst src, const Position & /*position*/) {
 }
 
 bool canConvertFromJson(JsonVariantConst src, const ENUVector & /*enu*/) {
-  return src["east"].is<double>() && src["north"].is<double>();
+  return src["east"].is<float>() && src["north"].is<float>();
 }
 
 bool canConvertFromJson(JsonVariantConst src,
